updated, adjusted - stable code structure
