+++ StreamerBot/BotClients/Twitch/TwitchBotLiveMonitorSvc.cs
@@ -3,11 +3,15 @@ using StreamerBot.Static;
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
+using System.Globalization;
 using System.Reflection;
+using System.Runtime.CompilerServices;
 
 using TwitchLib.Api;
 using TwitchLib.Api.Core;
 using TwitchLib.Api.Services;
+using TwitchLib.Api.Services.Events.LiveStreamMonitor;
 
 namespace StreamerBot.BotClients.Twitch
 {
@@ -39,6 +43,7 @@ namespace StreamerBot.BotClients.Twitch
             LiveStreamMonitor = new LiveStreamMonitorService(new TwitchAPI(null, null, apilive, null), (int)Math.Round(TwitchFrequencyLiveNotifyTime, 0));
             SetLiveMonitorChannels(new());
         }
+
         /// <summary>
         /// Adds and updates the channels to monitor for if the streamer goes live.
         /// </summary>
@@ -117,5 +122,127 @@ namespace StreamerBot.BotClients.Twitch
             LiveStreamMonitor = null;
             return base.ExitBot();
         }
+
+        #region MultiLive Bot
+
+        public static MultiUserLiveBot.Data.DataManager MultiLiveDataManager { get; private set; }
+
+        private const int maxlength = 8000;
+
+        public static string MultiLiveStatusLog { get; set; } = "";
+
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        public void MultiConnect()
+        {
+            MultiLiveDataManager = new();
+            IsMultiConnected = true;
+            NotifyPropertyChanged(nameof(MultiLiveDataManager));
+        }
+
+        public void MultiDisconnect()
+        {
+            StopMultiLive();
+            MultiLiveDataManager = null;
+            IsMultiConnected = false;
+            NotifyPropertyChanged(nameof(MultiLiveDataManager));
+        }
+
+        public void UpdateChannels()
+        {
+            if (IsMultiLiveBotActive)
+            {
+                MultiLiveDataManager.SaveData();
+                SetLiveMonitorChannels(MultiLiveDataManager.GetChannelNames());
+            }
+            else
+            {
+                SetLiveMonitorChannels(new());
+            }
+            // TODO: localize the multilive bot data
+            LogEntry(string.Format(CultureInfo.CurrentCulture, "MultiLive Bot started and monitoring {0} channels.", LiveStreamMonitor.ChannelsToMonitor.Count.ToString(CultureInfo.CurrentCulture)), DateTime.Now.ToLocalTime());
+        }
+
+        public void StartMultiLive()
+        {
+            if (IsMultiConnected && !IsMultiLiveBotActive)
+            {
+                IsMultiLiveBotActive = true;
+                UpdateChannels();
+            }
+        }
+
+        public void StopMultiLive()
+        {
+            if (IsMultiConnected && IsMultiLiveBotActive)
+            {
+                IsMultiLiveBotActive = false;
+                UpdateChannels();
+                LogEntry("MultiLive Bot stopped.", DateTime.Now.ToLocalTime());
+            }
+        }
+
+        public void SendMultiLiveMsg(OnStreamOnlineArgs e)
+        {
+            if (IsMultiLiveBotActive)
+            {
+                // true posted new event, false did not post
+                bool PostedLive = MultiLiveDataManager.PostStreamDate(e.Stream.UserName, e.Stream.StartedAt);
+
+                if (PostedLive)
+                {
+
+                    bool MultiLive = MultiLiveDataManager.CheckStreamDate(e.Channel, e.Stream.StartedAt);
+
+                    if ((OptionFlags.PostMultiLive && MultiLive) || !MultiLive)
+                    {
+                        // get message, set a default if otherwise deleted/unavailable
+                        string msg = OptionFlags.LiveMsg ?? "@everyone, #user is now live streaming #category - #title! Come join and say hi at: #url";
+
+                        // keys for exchanging codes for representative names
+                        Dictionary<string, string> dictionary = new()
+                        {
+                            { "#user", e.Stream.UserName },
+                            { "#category", e.Stream.GameName },
+                            { "#title", e.Stream.Title },
+                            { "#url", "https://www.twitch.tv/" + e.Stream.UserName }
+                        };
+
+                        LogEntry(VariableParser.ParseReplace(msg, dictionary), e.Stream.StartedAt);
+                        foreach (Tuple<string, Uri> u in MultiLiveDataManager.GetWebLinks())
+                        {
+                            if (u.Item1 == "Discord")
+                            {
+                                DiscordWebhook.SendMessage(u.Item2, VariableParser.ParseReplace(msg, dictionary));
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        private void NotifyPropertyChanged([CallerMemberName] string propertyName = "")
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        }
+
+        /// <summary>
+        /// Event to handle when the Twitch client sends and event. Updates the StatusLog property with the logged activity.
+        /// </summary>
+        /// <param name="data">The string of the message.</param>
+        /// <param name="dateTime">The time of the event.</param>
+        public void LogEntry(string data, DateTime dateTime)
+        {
+            if (MultiLiveStatusLog.Length + dateTime.ToString().Length + data.Length + 2 >= maxlength)
+            {
+                MultiLiveStatusLog = MultiLiveStatusLog[MultiLiveStatusLog.IndexOf('\n')..];
+            }
+
+            MultiLiveStatusLog += dateTime.ToString() + " " + data + "\n";
+
+            NotifyPropertyChanged(nameof(MultiLiveStatusLog));
+        }
+
+        #endregion
     }
 }
