+++ StreamerBot/StreamerBot.xaml
@@ -12,7 +12,6 @@
         <ResourceDictionary>
             <GUI:GUITwitchBots x:Name="TwitchBot" x:Key="TwitchBot" />
             <GUI:GUIDataManagerViews x:Name="DataViews" x:Key="DataViews" />
-            <GUI:GUIMultiLiveDataManager x:Name="MultiLiveDataManager" x:Key="MultiLiveDataManager" />
             <Static:MsgVarHelp x:Name="CommandInfo" x:Key="CommandInfo" />
 
             <Style TargetType="{x:Type DockPanel}">
@@ -453,7 +452,7 @@
                                 </DockPanel>
                             </ScrollViewer>
                         </TabItem>
-                        <TabItem Header="Variable Options">
+                        <TabItem Header="Variables">
                             <ListView DockPanel.Dock="Right" DataContext="{Binding Mode=OneWay, Source={StaticResource CommandInfo}}" ItemsSource="{Binding Collection}" >
                                 <ListView.View>
                                     <GridView>
@@ -525,7 +524,7 @@
                                     </Style>
                                 </Grid.Resources>
                                 <Grid.DataContext>
-                                    <Binding Mode="OneWay" Source="{StaticResource MultiLiveDataManager}"/>
+                                    <Binding Path="TwitchLiveMonitor" Source="{StaticResource TwitchBot}"/>
                                 </Grid.DataContext>
                                 <Grid.RowDefinitions>
                                     <RowDefinition Height="50" />
@@ -568,7 +567,7 @@
 
                                 <DockPanel x:Name="Panel_BotActivity" Grid.Row="1" Grid.Column="4" Visibility="Collapsed">
                                     <Label DockPanel.Dock="Top" Content="Log of Bot Activity:" />
-                                    <TextBox DockPanel.Dock="Bottom" x:Name="TB_BotActivityLog" Margin="5" TextWrapping="Wrap" Background="#FFEEEEEE" ScrollViewer.VerticalScrollBarVisibility="Auto" IsReadOnly="True" VerticalScrollBarVisibility="Visible" IsInactiveSelectionHighlightEnabled="True" TextChanged="TB_BotActivityLog_TextChanged" Text="{Binding MultiLiveStatusLog}" DataContext="{Binding MultiLiveDataManager, Mode=OneWay}" />
+                                    <TextBox DockPanel.Dock="Bottom" x:Name="TB_BotActivityLog" Margin="5" TextWrapping="Wrap" Background="#FFEEEEEE" ScrollViewer.VerticalScrollBarVisibility="Auto" IsReadOnly="True" VerticalScrollBarVisibility="Visible" IsInactiveSelectionHighlightEnabled="True" TextChanged="TB_BotActivityLog_TextChanged" Text="{Binding MultiLiveStatusLog}"  />
                                 </DockPanel>
                             </Grid>
                         </TabItem>
